+++ src/components/organization/Organization.js
@@ -15,7 +15,7 @@ import path from 'path';
 import { AuthenticationContext } from '..';
 
 import { Code } from '@material-ui/icons';
-import { apiPath, get } from '../../core';
+import { get, isSelectedOrgWritable} from '../../core';
 
 const useStyles = makeStyles((theme) => ({
   avatar: { borderRadius: '20%' },
@@ -57,40 +57,25 @@ function Organization({
   }, [url, organization, config, getData]);
 
   const _onOrganization = useCallback(async () => {
-
-    const teamsUrl = path.join(apiPath, 'orgs/'+org.username+'/teams');
-    let teamsResults = await get({ url: teamsUrl, config });
-    // if there are no teams, then go ahead and set it
-    if ( !teamsResults ) {
+    if ( org === undefined ) {
       onOrganization(org);
       return;
     }
 
-    // if there are teams, then ensure user has write access
     // first get the user (if logged in)
+    // if not logged in, go ahead and set the org
     let user = contextAuthentication.user.login;
     if ( !user ) {
       onOrganization(org);
       return;
     }
 
-    // https://git.door43.org/api/v1/teams/172/members/cecil.new?access_token=PlaygroundTesting
-    // test for membership, looking for something other than "read"
-    let permission = 'read';
-    for ( let j=0; j < teamsResults.length; j++ ) {
-      let isTeamMemberUrl = path.join(apiPath, 'teams/'+teamsResults[j].id+'/members/'+user);
-      let teamMemberResult = await get({url: isTeamMemberUrl, config});
-      if ( teamMemberResult === null ) {
-        // not a team member
-        continue;
-      }
-      let _permission = teamsResults[j].permission;
-      if ( _permission !== 'read' ) {
-        permission = 'write';
-        break
-      }
-    }
-    if ( permission === 'read' ) {
+    // if user does not have write access to the org
+    // i.e., is only in teams with read access
+    // then show alert and do not set the org
+    // make the user pick another
+    const hasWritePermissions = await isSelectedOrgWritable({org,user,config});
+    if ( ! hasWritePermissions ) {
       alert(
         "Your door43 user account doesn't have permission to edit the " +
         "files in the organization you have chosen.\n" +
