+++ src/core/gitea-api/users/orgs/orgs.ts
@@ -11,10 +11,38 @@ export interface Organization {
   username: string;
   visibility: string;
   website: string;
-  permission?: string;
 }
 
 export function getCurrentUserOrgs({ config }: { config: APIConfig }): Promise<Organization[]> {
   const url = path.join(apiPath, 'user/orgs');
   return get({ url, config });
+}
+
+export async function isSelectedOrgWritable({ org, user, config}: { org: Organization, user:string, config:APIConfig }): Promise<boolean> {
+
+  const teamsUrl = path.join(apiPath, 'orgs/'+org.username+'/teams');
+  let teamsResults = await get({ url: teamsUrl, config });
+  // if there are no teams, then skip other checks
+  if ( !teamsResults ) {
+    return true;
+  }
+
+  // https://git.door43.org/api/v1/teams/172/members/cecil.new?access_token=PlaygroundTesting
+  // test for membership, looking for something other than "read"
+  let permission = false;
+  for ( let j=0; j < teamsResults.length; j++ ) {
+    let isTeamMemberUrl = path.join(apiPath, 'teams/'+teamsResults[j].id+'/members/'+user);
+    let teamMemberResult = await get({url: isTeamMemberUrl, config});
+    if ( teamMemberResult === null ) {
+      // not a team member
+      continue;
+    }
+    let _permission = teamsResults[j].permission;
+    if ( _permission !== 'read' ) {
+      permission = true;
+      break;
+    }
+  }
+
+  return permission;
 }
